Update & Fix: My pagination is working fine and i figured out some bug that was messing my code

+++ apps/client/lib/gqlQueries.ts
@@ -1,7 +1,7 @@
 import gql from "graphql-tag";
 
 export const GET_POSTS = gql`
-  query posts($skip: Float, $take: Float) {
+  query posts($skip: Int, $take: Int) {
     posts(skip: $skip, take: $take) {
       id
       title
